+++ distributions/univariate/discrete/HyperGeometricRand.h
@@ -12,20 +12,20 @@
  */
 class RANDLIBSHARED_EXPORT HyperGeometricRand : public DiscreteDistribution
 {
-    int N = 1; ///< population size
-    int K = 1; /// number of possible successes
-    int n = 1; /// number of draws
+    size_t N = 1; ///< population size
+    size_t K = 1; /// number of possible successes
+    size_t n = 1; /// number of draws
     double pmfCoef = 1; ///< C(N, n)
     double p0 = 1; ///< K/N
 
 public:
-    HyperGeometricRand(int totalSize, int drawsNum, int successesNum);
+    HyperGeometricRand(size_t totalSize, size_t drawsNum, size_t successesNum);
     std::string Name() const override;
     SUPPORT_TYPE SupportType() const override { return FINITE_T; }
     int MinValue() const override { return std::max(0, n - N + K); }
     int MaxValue() const override { return std::min(n, K); }
 
-    void SetParameters(int totalSize, int drawsNum, int successesNum);
+    void SetParameters(size_t totalSize, size_t drawsNum, size_t successesNum);
     inline int GetTotalSize() { return N; }
     inline int GetDrawsNum() { return n; }
     inline int GetSuccessesNum() { return K; }
