+++ distributions/univariate/discrete/ZipfRand.h
@@ -13,7 +13,7 @@
 class RANDLIBSHARED_EXPORT ZipfRand : public DiscreteDistribution
 {
     double s = 1; ///< exponent
-    int n = 1; ///< number
+    size_t n = 1; ///< number
     double invHarmonicNumber = 1; /// 1 / H(s, n)
 
     static constexpr int tableSize = 16;
@@ -21,15 +21,15 @@ class RANDLIBSHARED_EXPORT ZipfRand : public DiscreteDistribution
     double table[tableSize];
 
 public:
-    ZipfRand(double exponent, int number);
+    ZipfRand(double exponent, size_t number);
     std::string Name() const override;
     SUPPORT_TYPE SupportType() const override { return FINITE_T; }
     int MinValue() const override { return 1; }
     int MaxValue() const override { return n; }
 
-    void SetParameters(double exponent, int number);
+    void SetParameters(double exponent, size_t number);
     inline double GetExponent() const { return s; }
-    inline int GetNumber() const { return n; }
+    inline size_t GetNumber() const { return n; }
 
     double P(const int & k) const override;
     double logP(const int & k) const override;
